# src/openapi/system-spec.yaml
openapi: 3.0.3
info:
  title: HAXcms System API
  description: Control-plane API for authentication, lifecycle/admin, discovery/meta, and system settings/configuration.
  version: 0.0.0
servers:
  - url: /
    description: System API base path
...
paths:
...
  /system/api/v1/sites:
    get:
      tags:
        - lifecycle
      operationId: listSites
      summary: List all sites available to the authenticated user
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: Site listing payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - lifecycle
      operationId: createSite
      summary: Create a new site
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Site creation response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/sites/{siteName}:
    parameters:
      - $ref: "#/components/parameters/SiteName"
    get:
      tags:
        - lifecycle
      operationId: siteInfoGet
      summary: Return metadata for a specific site
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: Site metadata response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - lifecycle
      operationId: siteInfoPost
      summary: Return metadata for a specific site
      security:
        - bearerAuth: []
          userTokenQuery: []
      responses:
        "200":
          description: Site metadata response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/sites/{siteName}/clone:
    parameters:
      - $ref: "#/components/parameters/SiteName"
    post:
      tags:
        - lifecycle
      operationId: cloneSite
      summary: Clone an existing site
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Site clone response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/sites/{siteName}/archive:
    parameters:
      - $ref: "#/components/parameters/SiteName"
    post:
      tags:
        - lifecycle
      operationId: archiveSite
      summary: Archive a site
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Site archive response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/sites/{siteName}/download:
    parameters:
      - $ref: "#/components/parameters/SiteName"
    post:
      tags:
        - lifecycle
      operationId: downloadSite
      summary: Generate and return a downloadable site archive reference
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Site download response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/sites/{siteName}/download-skeleton:
    parameters:
      - $ref: "#/components/parameters/SiteName"
    post:
      tags:
        - lifecycle
      operationId: downloadSiteSkeleton
      summary: Generate and return a reusable skeleton for a site
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Site skeleton response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/api/v1/sites/{siteName}/save-as-template:
    parameters:
      - $ref: "#/components/parameters/SiteName"
    post:
      tags:
        - lifecycle
      operationId: saveSiteAsTemplate
      summary: Save a site as a reusable system template
      security:
        - bearerAuth: []
          userTokenQuery: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteActionRequest"
      responses:
        "200":
          description: Save-as-template response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
    SiteName:
      in: path
      name: siteName
      required: true
      schema:
        type: string
      description: Machine name of the site
...
  responses:
    BadRequest:
      description: Request payload was invalid or incomplete
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Forbidden:
      description: Authenticated user is not authorized for this route
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    userTokenQuery:
      type: apiKey
      in: query
      name: user_token
    userTokenHeader:
      type: apiKey
      in: header
      name: X-HAXCMS-User-Token
    siteTokenQuery:
      type: apiKey
      in: query
      name: site_token
    appStoreTokenQuery:
      type: apiKey
      in: query
      name: appstore_token
  schemas:
    ApiEnvelope:
      type: object
      properties:
        status:
          type: integer
        data: {}
      additionalProperties: true
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            message:
              type: string
...
    SiteActionRequest:
      type: object
      properties:
        site:
          type: object
      additionalProperties: true
...